rename: chat 디렉토리 하위에 chat 디렉토리 추가해서 chatroom 디렉토리와 구분
