+++ chat/src/main/kotlin/kpring/chat/chat/model/Chat.kt
@@ -1,4 +1,4 @@
-package kpring.chat.model
+package kpring.chat.chat.model
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
