+++ chat/src/main/kotlin/kpring/chat/chat/repository/ChatRepository.kt
@@ -1,6 +1,6 @@
-package kpring.chat.repository
+package kpring.chat.chat.repository
 
-import kpring.chat.model.Chat
+import kpring.chat.chat.model.Chat
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.stereotype.Repository
