+++ chat/src/main/kotlin/kpring/chat/chat/service/ChatService.kt
@@ -1,8 +1,8 @@
-package kpring.chat.service
+package kpring.chat.chat.service
 
-import kpring.chat.model.Chat
-import kpring.chat.repository.ChatRepository
-import kpring.core.chat.dto.request.CreateChatRequest
+import kpring.chat.chat.model.Chat
+import kpring.chat.chat.repository.ChatRepository
+import kpring.core.chat.chat.dto.request.CreateChatRequest
 import org.springframework.stereotype.Service
 
 @Service
