+++ core/src/main/kotlin/kpring/core/chat/chat/dto/request/CreateChatRequest.kt
@@ -1,4 +1,4 @@
-package kpring.core.chat.dto.request
+package kpring.core.chat.chat.dto.request
 
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
