+++ core/src/main/kotlin/kpring/core/chat/chat/dto/response/CreateChatResponse.kt
@@ -1,4 +1,4 @@
-package kpring.core.chat.dto.response
+package kpring.core.chat.chat.dto.response
 
 data class CreateChatResponse (
     val code: Int,
